reflection_agent/pyproject.toml
[tool.poetry]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = ["Sumeet Vyas <sumeet.vyas@msci.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.1"
black = "^25.9.0"
isort = "^7.0.0"
langchain = "^1.0.2"
langchain-openai = "^1.0.1"
langgraph = "^1.0.1"
grandalf = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
